наладил эмуляцию мыши

+++ ESP8266WebServer/src/controllers/Mouse.h
@@ -10,17 +10,17 @@ namespace Mouse{
     ClientProcessing cpr;
     SerialModule serial;
     JsonDocument parsed;
-    
+
     void set(){
         if(cpr.CheckCLientReq(&parsed, HTTP_POST)){
             uint8 data[7] = {};
             data[0] = MOUSE;
             data[1] = SET_COMMAND;
             data[2] = 4;
-            data[3] = parsed["buttons"];
-            data[4] = parsed["move_x"];
-            data[5] = parsed["move_y"];
-            data[6] = parsed["scroll"];   
+            data[3] = (uint8_t)parsed["buttons"];
+            data[4] = (uint8_t)parsed["move_x"];
+            data[5] = (uint8_t)parsed["move_y"];
+            data[6] = (uint8_t)parsed["scroll"];   
 
             cpr.AnswerClient(serial.SendDataWithWait(data, 7));
         }
@@ -41,7 +41,7 @@ namespace Mouse{
             data[0] = MOUSE;
             data[1] = CLICK_COMMAND;
             data[2] = 1; 
-            data[3] = parsed["buttons"];
+            data[3] = (uint8_t)parsed["buttons"];
 
             cpr.AnswerClient(serial.SendDataWithWait(data, 4));
         }
